continued iteration on database layout and controller flows

+++ src/main/java/space/knightdev/dghandicap/dao/course/CourseLayout.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class CourseLayout {
     private int layoutId;
     private String name;
-    private int ssa;
+    private double ssa;
 }
